setup db and etl
