+++ .gitignore
@@ -1,3 +1,4 @@
 node_modules
 .DS_Store
 .env
+server/tmp/*.csv
