+++ package.json
@@ -19,7 +19,7 @@
   },
   "scripts": {
     "start": "concurrently 'react-scripts start' 'cd server && npm run start:dev'",
-    "start:ui": "react-scripts start",
+    "start-ui": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
