+++ server/package.json
@@ -5,7 +5,10 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start:dev": "nodemon index.js"
+    "start-dev": "nodemon index.js",
+    "migrate": "npm run parse-dot-env node-pg-migrate",
+    "migrate:no-dot-env": "node-pg-migrate",
+    "parse-dot-env": "eval $(egrep -v '^#' .env | xargs)"
   },
   "author": "",
   "license": "ISC",
