+++ server/sample.env
@@ -1,0 +1,1 @@
+DATABASE_URL=postgres://postgres:postgres@localhost:54321/postgres
